requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-storages[boto3]>=1.13.0  
boto3>=1.17.0
idna==3.10
python-dotenv==1.0.1
mpmath==1.3.0
requests==2.32.3
sqlparse==0.5.3
sympy==1.13.3
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.8.2
psycopg2==2.9.10
gunicorn==23.0.0
weasyprint